(refactor) mv payment.resource.ts to mpesa-resource.tsx
