+++ packages/esm-billing-app/src/invoice/payments/initiate-payment/initiate-payment.component.tsx
@@ -4,13 +4,13 @@ import { Button, Form, ModalBody, ModalHeader, TextInput, Layer, InlineNotificat
 import styles from './initiate-payment.scss';
 import { Controller, useForm } from 'react-hook-form';
 import { MappedBill } from '../../../types';
-import { initiateStkPush } from '../payment.resource';
 import { showSnackbar, useConfig } from '@openmrs/esm-framework';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { formatPhoneNumber } from '../utils';
 import { Buffer } from 'buffer';
 import { useSystemSetting } from '../../../hooks/getMflCode';
+import { initiateStkPush } from '../../../m-pesa/mpesa-resource';
 
 const InitiatePaymentSchema = z.object({
   phoneNumber: z
