+++ packages/esm-billing-app/src/m-pesa/mpesa-resource.tsx
@@ -14,7 +14,7 @@ const swrOption = {
   errorRetryCount: 2,
 };
 
-export const usePaymentModes = (excludeWaiver: boolean = true) => {
+export const usePaymentModes = () => {
   const { excludedPaymentMode } = useConfig<BillingConfig>();
   const url = `/ws/rest/v1/cashier/paymentMode`;
   const { data, isLoading, error, mutate } = useSWR<{ data: { results: Array<PaymentMethod> } }>(
@@ -27,14 +27,14 @@ export const usePaymentModes = (excludeWaiver: boolean = true) => {
       ? data?.data?.results.filter((mode) => !excludedPaymentMode.some((excluded) => excluded.uuid === mode.uuid)) ?? []
       : data?.data?.results ?? [];
   return {
-    paymentModes: excludeWaiver ? allowedPaymentModes : data?.data?.results,
+    paymentModes: allowedPaymentModes,
     isLoading,
     mutate,
     error,
   };
 };
 
-export const generateStkAccessToken = async (authorizationUrl: string) => {
+export const generateStkAccessToken = async (authorizationUrl: string, setNotification) => {
   try {
     const consumerKey = '';
     const consumerSecret = '';
@@ -47,13 +47,14 @@ export const generateStkAccessToken = async (authorizationUrl: string) => {
     const { access_token } = await response.json();
     return access_token;
   } catch (error) {
+    setNotification('Unable to reach the MPESA server, please try again later.');
     throw error;
   }
 };
 
-export const initiateStkPush = async (payload, initiateUrl: string, authorizationUrl: string) => {
+export const initiateStkPush = async (payload, initiateUrl: string, authorizationUrl: string, setNotification) => {
   try {
-    const access_token = await generateStkAccessToken(authorizationUrl);
+    const access_token = await generateStkAccessToken(authorizationUrl, setNotification);
     const headers = {
       'Content-Type': 'application/json',
       Authorization: `Bearer ${access_token}`,
@@ -63,8 +64,10 @@ export const initiateStkPush = async (payload, initiateUrl: string, authorizatio
       headers: headers,
       body: JSON.stringify(payload),
     });
+
     return await response.json();
   } catch (err) {
+    setNotification('Unable to initiate Lipa Na Mpesa, please try again later.');
     throw err;
   }
 };
